+++ doc/source/changelog.rst
@@ -2,6 +2,59 @@
 Changelog
 #########
 
+Version 4.1 (2020-05-27)
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+- feat: add "must_link_method" option to "pool" clustering
+
+Version 4.0 (2020-05-15)
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+- feat: add support for cannot/must link constraints in "pool" clustering
+- BREAKING: make one_hot_encoding return SlidingWindowFeature with labels
+- feat: add "labels" optional attribute to SlidingWindowFeature
+
+Version 3.7.1 (2020-03-23)
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+- fix: fix corner case in pyannote.core.utils.numpy.one_hot_encoding
+
+Version 3.7 (2020-03-21)
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+- feat: add new method SlidingWindowFeature.align
+
+Version 3.6 (2020-03-03)
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+- feat: add collar argument to Timeline.support (@nryant)
+- feat: add new method Timeline.covers (@PaulLerner)
+- fix: check for spaces before writing with RTTM or UEM format (@PaulLerner)
+- improve: speed up Annotation.{from_df | from_json} (@nryant)
+
+Version 3.5 (2020-01-23)
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+- feat: add align_last option to SlidingWindow.__call__
+
+Version 3.4 (2020-01-21)
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+- feat: add SlidingWindow.__call__ to slide over a specific support
+- feat: add pyannote.core.utils.random
+
+Version 3.3 (2020-01-08)
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+- BREAKING: remove pyannote.core.Scores
+- fix: fix legend corner case with empty Annotation instances
+- fix: fix {Timeline|Annotation}.crop with overlapping segments in support
+
+Version 3.2.2 (2020-01-06)
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+- fix: fix Segment.__bool__ return type
+
 Version 3.2 (2019-12-13)
 ~~~~~~~~~~~~~~~~~~~~~~~~